+++ packages/pigeon/run_tests.sh
@@ -88,24 +88,16 @@ test_pigeon_android() {
 # Compiles the pigeon file to a temp directory and attempts to run the dart
 # analyzer on it.
 test_pigeon_dart() {
-  echo "test_pigeon_dart($1)"
-  temp_dir=$(mktmpdir)
+  echo "test_pigeon_dart($1, $2)"
+  local flutter_project_dir=$2
 
   $run_pigeon \
     --input $1 \
-    --dart_out $temp_dir/pigeon.dart &
-  gen_pid=$!
+    --dart_out $flutter_project_dir/lib/pigeon.dart
 
-  wait $gen_pid
+  dart analyze $flutter_project_dir/lib/pigeon.dart --fatal-infos --fatal-warnings
 
-  # `./e2e_tests/test_objc/.packages` is used to get access to Flutter since
-  # Pigeon doesn't depend on Flutter.
-  dart analyze $temp_dir/pigeon.dart --fatal-infos --fatal-warnings --packages ./e2e_tests/test_objc/.packages &
-  analyze_pid=$!
-
-  wait $analyze_pid
-
-  rm -rf $temp_dir
+  rm $flutter_project_dir/lib/pigeon.dart
 }
 
 print_usage() {
@@ -200,23 +192,22 @@ run_mock_handler_tests() {
 }
 
 run_dart_compilation_tests() {
-  # DEPRECATED: These tests are deprecated, use run_flutter_unittests instead.
-  # Make sure the artifacts are present.
-  flutter precache
-  # Make sure flutter dependencies are available.
-  pushd $PWD
-  cd e2e_tests/test_objc/
-  flutter pub get
-  popd
-  test_pigeon_dart ./pigeons/all_void.dart
-  test_pigeon_dart ./pigeons/async_handlers.dart
-  test_pigeon_dart ./pigeons/host2flutter.dart
-  test_pigeon_dart ./pigeons/list.dart
-  test_pigeon_dart ./pigeons/message.dart
-  test_pigeon_dart ./pigeons/void_arg_flutter.dart
-  test_pigeon_dart ./pigeons/void_arg_host.dart
-  test_pigeon_dart ./pigeons/voidflutter.dart
-  test_pigeon_dart ./pigeons/voidhost.dart
+  local temp_dir=$(mktmpdir)
+  local flutter_project_dir=$temp_dir/project
+
+  flutter create --platforms="android" $flutter_project_dir 1> /dev/null
+
+  test_pigeon_dart ./pigeons/all_void.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/async_handlers.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/host2flutter.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/list.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/message.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/void_arg_flutter.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/void_arg_host.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/voidflutter.dart $flutter_project_dir
+  test_pigeon_dart ./pigeons/voidhost.dart $flutter_project_dir
+
+  rm -rf $temp_dir
 }
 
 run_ios_unittests() {
@@ -225,6 +216,7 @@ run_ios_unittests() {
   gen_ios_unittests_code ./pigeons/async_handlers.dart ""
   gen_ios_unittests_code ./pigeons/background_platform_channels.dart "BC"
   gen_ios_unittests_code ./pigeons/enum.dart "AC"
+  gen_ios_unittests_code ./pigeons/enum_args.dart "EA"
   gen_ios_unittests_code ./pigeons/host2flutter.dart ""
   gen_ios_unittests_code ./pigeons/list.dart "LST"
   gen_ios_unittests_code ./pigeons/message.dart ""
@@ -286,6 +278,7 @@ run_android_unittests() {
   gen_android_unittests_code ./pigeons/async_handlers.dart AsyncHandlers
   gen_android_unittests_code ./pigeons/background_platform_channels.dart BackgroundPlatformChannels
   gen_android_unittests_code ./pigeons/enum.dart Enum
+  gen_android_unittests_code ./pigeons/enum_args.dart EnumArgs
   gen_android_unittests_code ./pigeons/host2flutter.dart Host2Flutter
   gen_android_unittests_code ./pigeons/java_double_host_api.dart JavaDoubleHostApi
   gen_android_unittests_code ./pigeons/list.dart PigeonList