+++ packages/cross_file/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.3+1
+
+* Fixes `lastModified` unimplemented error description.
+
 ## 0.3.3
 
 * Removes unused Flutter dependencies.