+++ packages/flutter_template_images/CHANGELOG.md
@@ -1,6 +1,12 @@
-## NEXT
+## 4.1.1
+
+* Removes empty Dart file.
+* Opts in to NNBD (which is a no-op) so it's not flagged by pub.dev.
+
+## 4.1.0
 
 * Updates package description.
+* Adds macOS template app icons, updated to Big Sur style.
 
 ## 4.0.0