+++ packages/go_router_builder/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.5
+
+- Update example to avoid using `push()` to push the same page since is not supported. [#105150](https://github.com/flutter/flutter/issues/105150)
+
+## 1.0.4
+
+- Adds `push` method to generated GoRouteData's extension. [#103025](https://github.com/flutter/flutter/issues/103025)
+
 ## 1.0.3
 
 - Fixes incorrect separator at location path on Windows. [#102710](https://github.com/flutter/flutter/issues/102710)